move common utils to library
