+++ src/renderer/library/styles/StatsFooter.ts
