+++ src/renderer/library/styles/index.ts
@@ -1,4 +1,5 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export * from './helpers';
+export * from './ScrollWrappers';
+export * from './StatsFooter';
