+++ src/renderer/screens/Import/Ledger/Manage.tsx
@@ -16,7 +16,7 @@ import { getSortedLocalLedgerAddresses } from '@/renderer/utils/ImportUtils';
 import { useAddresses } from '@/renderer/contexts/import/Addresses';
 import { useState } from 'react';
 import { ButtonPrimaryInvert } from '@/renderer/kits/Buttons/ButtonPrimaryInvert';
-import { Scrollable, StatsFooter } from '@/renderer/library/utils';
+import { Scrollable, StatsFooter } from '@/renderer/library/styles';
 import { ControlsWrapper, SortControlLabel } from '@/renderer/utils/common';
 import type { ImportLedgerManageProps } from '../types';
