+++ src/renderer/screens/OpenGov/Tracks/index.tsx
@@ -16,12 +16,9 @@ import { ContentWrapper } from '@app/screens/Wrappers';
 import { StickyHeadings, TrackGroup } from './Wrappers';
 import { ButtonPrimaryInvert } from '@/renderer/kits/Buttons/ButtonPrimaryInvert';
 import { TrackRow } from './TrackRow';
-import { Scrollable, StatsFooter } from '@/renderer/library/utils';
-import {
-  renderPlaceholders,
-  ControlsWrapper,
-  SortControlButton,
-} from '@/renderer/utils/common';
+import { Scrollable, StatsFooter } from '@/renderer/library/styles';
+import { renderPlaceholders } from '@/renderer/library/utils';
+import { ControlsWrapper, SortControlButton } from '@/renderer/utils/common';
 import type { HelpItemKey } from '@/renderer/contexts/common/Help/types';
 import type { TracksProps } from '../types';
